requirements.txt
flask==3.0.3
flask-restful==0.3.10
pyyaml==6.0.2